by pasupathi

+++ Master/AirlineMangementsystem/src/main/java/com/dxc/airline/repository/AdminRepository.java
@@ -1,10 +1,9 @@
 package com.dxc.airline.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 
 import com.dxc.airline.model.Admin;
 
-public interface AdminRepository extends JpaRepository<Admin, Integer> {
+public interface AdminRepository extends JpaRepository<Admin, Long> {
 
 }
